+++ server/src/managers/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-    CommandManager: require('./commands'),
-    Config: require('./config'),
-    ImportsManager: require('./imports-manager'),
-    Stats: require('./stats')
-};