+++ server/src/scripts/configure.js
@@ -1,5 +1,5 @@
 /**
-  * Server configuration script, used reconfiguring server options
+  * Server configuration script, to (re)configure server options
   *
   * Version: v2.0.0
   * Developer: Marzavec ( https://github.com/marzavec )
@@ -11,13 +11,11 @@
 
 // import required classes
 const path = require('path');
-const ImportsManager = require('../managers/imports-manager');
-const ConfigManager = require('../managers/config');
+const ConfigManager = require('../serverLib/ConfigManager');
+const SetupWizard = require('./configLib/SetupWizard');
 
 // import and initialize configManager & dependencies
-const importManager = new ImportsManager(null, path.join(__dirname, '../..'));
-importManager.init();
-const configManager = new ConfigManager(null, path.join(__dirname, '../..'), importManager);
+const serverConfig = new ConfigManager(path.join(__dirname, '../..'));
+const setup = new SetupWizard(serverConfig);
 
-// execute config load with `reconfiguring` flag set to true
-configManager.load(true);
+setup.start();